+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
 
 - ci(): Add some prebuilt fabric in the dist folder [#10178](https://github.com/fabricjs/fabric.js/pull/10178)
+- chore(): Add more generic font families to FabricText.genericFonts [#10167](https://github.com/fabricjs/fabric.js/pull/10167)
 
 ## [6.4.2]